+++ lib/SILOptimizer/Mandatory/DIMemoryUseCollectorOwnership.cpp
@@ -13,6 +13,7 @@
 #define DEBUG_TYPE "definite-init"
 #include "DIMemoryUseCollectorOwnership.h"
 #include "swift/AST/Expr.h"
+#include "swift/SIL/ApplySite.h"
 #include "swift/SIL/InstructionUtils.h"
 #include "swift/SIL/SILArgument.h"
 #include "swift/SIL/SILBuilder.h"
